pyproject.toml
[tool.poetry]
name = "online-learning-simple-demo"
version = "0.1.0"
description = ""
authors = ["Adrian Zuur <adrianzuur@gmail.com>"]
readme = "README.md"
packages = [{include = "online_learning_simple_demo"}]

[tool.poetry.dependencies]
python = "^3.12"
langchain = "^0.2.7"
langchain-openai = "^0.1.16"
scikit-learn = "^1.5.1"
pandas = "^2.2.2"
python-dotenv = "^1.0.1"
tqdm = "^4.66.4"
pyarrow = "^16.1.0"
pydantic = "^2.8.2"
pydantic-settings = "^2.3.4"
pgvector = "^0.3.1"
psycopg = {extras = ["binary"], version = "^3.2.1"}
fastapi = "^0.111.1"
dash = "^2.17.1"
jinja2 = "^3.1.4"

[tool.poetry.group.dev.dependencies]
ruff = "^0.5.1"
black = "^24.4.2"
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
